+++ pom.xml
@@ -28,6 +28,7 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
+            <version>3.4.4</version>
         </dependency>
         <dependency>
             <groupId>com.h2database</groupId>
