+++ src/main/java/edens/zac/portfolio/backend/controller/dev/BlogControllerDev.java
@@ -1,4 +1,4 @@
-package edens.zac.portfolio.backend.controller.write;
+package edens.zac.portfolio.backend.controller.dev;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import edens.zac.portfolio.backend.model.BlogCreateDTO;
@@ -22,10 +22,10 @@ import java.util.List;
 @Slf4j
 @AllArgsConstructor
 @RestController
-@RequestMapping("/api/v1/blog")
+@RequestMapping("/api/write/blog")
 @Configuration
 @Profile("dev")
-public class BlogControllerWrite {
+public class BlogControllerDev {
 
     private final BlogService blogService;
